fix: postgresql query,insert macro and base bean create_time,update_time to DateTime(PG:timestraptz)

+++ tooling/deploy/migrations/sqlite/v20240710-1/sys.init.ddl.sql
@@ -21,7 +21,7 @@
 --
 -- Create the users table.
 CREATE TABLE IF NOT EXISTS users (
-    id INTEGER PRIMARY KEY NOT NULL,
+    id BIGINT PRIMARY KEY NOT NULL,
     name VARCHAR(64) NULL,
     -- "账号昵称"
     email VARCHAR(64) NULL,
